V2

+++ database/migrations/2023_06_14_064329_create_output_voucher_items_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
+        //php artisan migrate:refresh --path=database/migrations/2023_06_14_064329_create_output_voucher_items_table.php
+
         Schema::create('output_voucher_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('output_voucher_id')->constrained()->onUpdate('cascade')->onDelete('cascade');
@@ -18,7 +20,7 @@ return new class extends Migration
             $table->foreignId('input_voucher_item_id')->constrained()->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('employee_id')->constrained()->onUpdate('cascade')->onDelete('cascade');
             $table->integer('count')->nullable();
-            $table->integer('price')->nullable();
+            $table->double('price')->nullable();
             $table->bigInteger('value')->nullable();
             $table->text('notes')->nullable();
             $table->softDeletes();
